port changes

+++ server/services/cart/controllers/cartController.js
@@ -3,8 +3,7 @@ const { Cart, sequelize, Food } = require("../models")
 class CartController {
   static async create (req, res, next){
     try {
-      // const UserId  = req.userData.id
-      const UserId = 1 // for testing purpose delete before production
+      const UserId  = req.headers.user_id
       const { FoodId, quantity, status } = req.body
 
       let cartFromDb = await Cart.findOne({where: {UserId, FoodId}, include: [ Food ]}) // Add include: Food | if Food model completed
@@ -19,13 +18,26 @@ class CartController {
           }
           next(err)
         } else {
-          cartFromDb = await Cart.update({quantity: totalQuantity}, {where: {UserId, FoodId}})
-          return res.status(201).json({cartFromDb, message: "Successfully adding item to cart"})
+          let data = await Cart.update({quantity: totalQuantity}, {where: {UserId, FoodId}, returning: true})
+          let cartAfterUpdate= data[1][0].dataValues
+          return res.status(200).json({
+            id: cartAfterUpdate.id,
+            UserId: cartAfterUpdate.UserId,
+            FoodId: cartAfterUpdate.FoodId,
+            quantity: cartAfterUpdate.quantity,
+            status: cartAfterUpdate.status 
+          })
         }
       } else{
         
         const newCart = await Cart.create({UserId, FoodId, quantity, status})
-        return res.status(201).json({cartFromDb, message: "Successfully adding item to cart"})
+        return res.status(201).json({
+          id: newCart.id,
+          UserId: newCart.UserId,
+          FoodId: newCart.FoodId,
+          quantity: newCart.quantity,
+          status: newCart.status 
+        })
       }
       
     } catch (error) {
@@ -33,12 +45,12 @@ class CartController {
     }
   }
 
-  static async getCarts(req, rest, next){
+  static async getCarts(req, res, next){
     try {
-      let UserId  = req.userData.id
+      let UserId  = req.headers.user_id
+      console.log(req.headers);
+      let carts = await Cart.findAll({where: {UserId}, include: [Food]}) // Add include: Food | if Food model completed
       
-      let carts = Cart.findAll({where: UserId}) // Add include: Food | if Food model completed
-
       return res.status(200).json({carts})
     } catch (error) {
       next(error)
@@ -47,12 +59,19 @@ class CartController {
 
   static async updateQuantity(req,res,next){
     try{
+      let UserId  = req.headers.user_id
       let { id } = req.params //Cart Id
       let { quantity } = req.body
 
-      await Cart.update({quantity}, {where: {id}})
-
-      return res.status(200).json({message: 'Success updating cart'})
+      let data= await Cart.update({quantity}, {where: {id, UserId}, returning: true})
+      let cartAfterUpdate= data[1][0].dataValues
+      return res.status(200).json({
+        id: cartAfterUpdate.id,
+        UserId: cartAfterUpdate.UserId,
+        FoodId: cartAfterUpdate.FoodId,
+        quantity: cartAfterUpdate.quantity,
+        status: cartAfterUpdate.status 
+      })
     }catch(err){
       next(err)
     }
@@ -60,9 +79,10 @@ class CartController {
 
   static async deleteCart(req,res,next){
     try{
+      let UserId  = req.headers.user_id
       let {id} = req.params //Cart Id
 
-      await Cart.destroy({where:{id}})
+      await Cart.destroy({where:{id, UserId}})
 
       return res.status(200).json({message: 'Cart successfully deleted'})
